# Conflicts:
#	.gitignore
#	.idea/php.xml
#	.idea/sqldialects.xml
#	apps/core/GestorErrores.php
#	apps/core/PgTimestamp.php
#	apps/encargossacd/view/encargo_select.html.twig
#	apps/notas/controller/acta_imprimir_mpdf.php
#	apps/notas/controller/acta_update.php
#	apps/notas/model/Select1011.php
#	apps/ubis/model/entity/GestorDelegacion.php
#	libs/vendor/composer/installed.php
#	libs/vendor/myclabs/deep-copy/.github/workflows/ci.yaml

+++ apps/certificados/controller/certificado_update.php
@@ -5,6 +5,7 @@
 use certificados\domain\entity\Certificado;
 use certificados\domain\repositories\CertificadoRepository;
 use core\ConfigGlobal;
+use core\ServerConf;
 use personas\model\entity\Persona;
 use web\DateTimeLocal;
 use function core\is_true;
@@ -25,17 +26,20 @@ $Qnom = (string)filter_input(INPUT_POST, 'nom');
 $Qidioma = (string)filter_input(INPUT_POST, 'idioma');
 $Qdestino = (string)filter_input(INPUT_POST, 'destino');
 $Qcertificado = (string)filter_input(INPUT_POST, 'certificado');
-$Qcopia = (string)filter_input(INPUT_POST, 'copia');
+$Qfirmado = (string)filter_input(INPUT_POST, 'firmado');
 $Qf_certificado = (string)filter_input(INPUT_POST, 'f_certificado');
+$Qf_enviado = (string)filter_input(INPUT_POST, 'f_enviado');
 
 $Qcertificado_old = (string)filter_input(INPUT_POST, 'certificado_old');
 
 /* convertir las fechas a DateTimeLocal */
 $oF_certificado = DateTimeLocal::createFromLocal($Qf_certificado);
+$oF_enviado = DateTimeLocal::createFromLocal($Qf_enviado);
 
 $error_txt = '';
 
 $certificadoRepository = new CertificadoRepository();
+
 if (is_true($Qnuevo)) {
     $Qid_item = $certificadoRepository->getNewId_item();
     $oCertificado = new Certificado();
@@ -56,14 +60,17 @@ $oCertificado->setNom($Qnom);
 $oCertificado->setIdioma($Qidioma);
 $oCertificado->setDestino($Qdestino);
 $oCertificado->setCertificado($Qcertificado);
-if (is_true($Qcopia)) {
-    $copia = TRUE;
+if (is_true($Qfirmado)) {
+    $firmado = TRUE;
 } else {
-    $copia = FALSE;
+    $firmado = FALSE;
 }
-$oCertificado->setCopia($copia);
-$oCertificado->setPropio(FALSE);
+$oCertificado->setFirmado($firmado);
+$oCertificado->setEsquema_emisor(ConfigGlobal::mi_region_dl());
 $oCertificado->setF_certificado($oF_certificado);
+if (!empty($oF_enviado)) {
+    $oCertificado->setF_enviado($oF_enviado);
+}
 
 if ($certificadoRepository->Guardar($oCertificado) === FALSE) {
     $error_txt .= $certificadoRepository->getErrorTxt();
@@ -72,8 +79,10 @@ if ($certificadoRepository->Guardar($oCertificado) === FALSE) {
 if (!empty($Qcertificado_old)) {
     $filename_sin_barra = str_replace('/', '_', $Qcertificado_old);
     $filename_sin_espacio = str_replace(' ', '_', $filename_sin_barra);
-    $filename_pdf = ConfigGlobal::DIR . '/log/tmp/' . $filename_sin_espacio . '.pdf';
-    unlink($filename_pdf);
+    $filename_pdf = ServerConf::DIR . '/log/tmp/' . $filename_sin_espacio . '.pdf';
+    if (is_file($filename_pdf)) {
+        unlink($filename_pdf);
+    }
 }
 
 if (empty($error_txt)) {
